requirements.txt
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
yt-dlp>=2023.11.16
gTTS>=2.3.2
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0